+++ src/main/java/com/petitapetit/miml/global/config/AsyncConfig.java
@@ -1,0 +1,40 @@
+package com.petitapetit.miml.global.config;
+
+import java.util.concurrent.Executor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.TaskDecorator;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+@Slf4j
+@Configuration
+@EnableAsync
+public class AsyncConfig {
+    @Bean
+    public Executor asyncTaskExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(4);
+        executor.setMaxPoolSize(4);
+        executor.setQueueCapacity(500);
+        executor.setThreadNamePrefix("Notification-");
+        executor.initialize();
+
+        executor.setTaskDecorator(new TaskDecorator() {
+            @Override
+            public Runnable decorate(Runnable runnable) {
+                return () -> {
+                    String threadName = Thread.currentThread().getName();
+                    long start = System.currentTimeMillis();
+                    log.info("Starting async task in thread : {}", threadName);
+                    runnable.run();
+                    long end = System.currentTimeMillis();
+                    log.info("It took "+(end-start)+"ms to process the async task.");
+                };
+            }
+        });
+
+        return executor;
+    }
+}
